Move PersonListItem.kt in the model/view package. Use the data class modifier for PersonItem class.
