+++ app/src/main/java/com/davidread/starwarsdatabase/model/view/PersonListItem.kt
@@ -1,4 +1,4 @@
-package com.davidread.starwarsdatabase.model
+package com.davidread.starwarsdatabase.model.view
 
 /**
  * Represents a list item that appears in the people list.
@@ -11,7 +11,7 @@ sealed class PersonListItem {
      * @param id Unique identifier for the person from the API.
      * @param name Person's name.
      */
-    class PersonItem(val id: Int, val name: String) : PersonListItem()
+    data class PersonItem(val id: Int, val name: String) : PersonListItem()
 
     /**
      * Represents a list item for a loading view.
