+++ app/src/main/java/com/davidread/starwarsdatabase/viewmodel/PeopleListFragmentViewModel.kt
@@ -1,7 +1,7 @@
 package com.davidread.starwarsdatabase.viewmodel
 
 import androidx.lifecycle.LiveData
-import com.davidread.starwarsdatabase.model.PersonListItem
+import com.davidread.starwarsdatabase.model.view.PersonListItem
 
 /**
  * Defines the structure of [PeopleListFragmentViewModelImpl].
